Move stories to components folder
