+++ .storybook/main.js
@@ -1,4 +1,4 @@
 module.exports = {
-  stories: ['./stories/**/*.stories.?(ts|tsx|js|jsx)'],
+  stories: ['../components/**/*.stories.?(ts|tsx|js|jsx)'],
   addons: ['@storybook/addon-ondevice-controls', '@storybook/addon-ondevice-actions'],
 };
