+++ components/leaf/OdysseyButton.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta } from '@storybook/react-native';
-import OdysseyButton from '../../../components/leaf/OdysseyButton';
+import OdysseyButton from './OdysseyButton';
 import { View } from 'react-native';
 
 const OdysseyButtonMeta: Meta<typeof OdysseyButton> = {
